Prefer ULID::MonotonicGenerator.new rather than ULID.monotonic_generate
Resolves #32
