+++ Steepfile
@@ -3,5 +3,5 @@ target :lib do
 
   check 'lib'
 
-  library 'securerandom', 'singleton'
+  library 'securerandom'
 end
